<FIX>Goal 모듈 비즈니스 로직 이관 #63
