+++ src/user/user.controller.ts
@@ -19,17 +19,12 @@ import { AuthGuard } from '@nestjs/passport';
 import { AuthService } from '../auth/auth.service';
 import { UserService } from './user.service';
 import { UserGoalService } from '../usergoal/userGoal.service';
-import { GoalService } from 'src/goal/goal.service';
 import { BadgeService } from 'src/badges/badge.service';
-import { BalanceService } from 'src/balances/balances.service';
 import { NaverAuthGuard } from '../auth/naver/naver-auth.guard';
 import { KakaoAuthGuard } from '../auth/kakao/kakao-auth.guard';
 import { GoogleOauthGuard } from '../auth/google/google-oauth.guard';
-import { AccessUserGoalDTO } from 'src/usergoal/dto/accessUserGoals.dto';
-import { ExitUserDTO } from './dto/exitUser.dto';
 import { UpdatePinCodeDTO } from './dto/updatePinCode.dto';
 import { ModifyUserInfoDTO } from './dto/modifyUser.dto';
-import { AccountsService } from 'src/accounts/accounts.service';
 import { User } from 'src/common/decorators/user.decorator';
 import { ConfigService } from '@nestjs/config';
 import { createHash } from 'crypto';
